requirements.txt
asgiref==3.6.0
Django==4.2.1
django-debug-toolbar==4.0.0
django-enumfields==2.1.1
django-environ==0.10.0
psycopg2==2.9.6
sqlparse==0.4.4
